find_optimised_snps

find_optimised_snps is used to find optimised SNPs set.

find_optimised_snps(
  seqc,
  metric = "simpson",
  goi = c(),
  accept_multiallelic = TRUE,
  number_of_result = 1,
  max_depth = 1,
  included_positions = c(),
  excluded_positions = c(),
  iterate_included = FALSE,
  bp = SerialParam(),
  ...
)

Arguments

seqc
list of sequences, either passed directly from process_allele or read_fasta or equivalence

metric
either `simpson` or `percent`

goi
group of interest, if creteria is percent, must be specified, ignored otherwise

accept_multiallelic
whether include positions with > 1 state in goi

number_of_result
number of results to return, 0 will be coerced to 1

max_depth
maximum depth to go before terminating, 0 means it will only calculate the metric for included position

included_positions
included positions

excluded_positions
excluded positions

iterate_included
whether to calculate index at each level of the included SNPs

bp
BiocParallel backend. Rule of thumbs: use MulticoreParam()

...
other parameters as needed

Value

Will return the resolution-optimised SNPs set, based on the metric.